+++ .github/tests/application/tests/TestCases/PluginIsActive/AdminPagesTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Korobochkin\BBPressPermalinksWithIdTestsApplication\Tests\TestCases\PluginIsActive;
 
 use Korobochkin\BBPressPermalinksWithIdTestsApplication\Tests\Abstracts\AbstractAdminPagesTest;
-use Korobochkin\BBPressPermalinksWithIdTestsApplication\Tests\TestCases\Common\ActivatePluginTest;
+use Korobochkin\BBPressPermalinksWithIdTestsApplication\Tests\TestCases\PluginIsNotActive\ActivatePluginTest;
 use PHPUnit\Framework\Attributes;
 
 /**
