+++ .github/tests/application/tests/TestCases/PluginIsActive/ForumsPageTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Korobochkin\BBPressPermalinksWithIdTestsApplication\Tests\TestCases\PluginIsActive;
 
 use Korobochkin\BBPressPermalinksWithIdTestsApplication\Tests\Abstracts\AbstractForumsPageTest;
-use Korobochkin\BBPressPermalinksWithIdTestsApplication\Tests\TestCases\Common\ActivatePluginTest;
+use Korobochkin\BBPressPermalinksWithIdTestsApplication\Tests\TestCases\PluginIsNotActive\ActivatePluginTest;
 use PHPUnit\Framework\Attributes;
 
 /**
